+++ src/MixerWidgets.hpp
@@ -598,6 +598,10 @@ struct TrackDisplay : EditableDisplayBase {
 			trackLPFAdjustSlider->box.size.x = 200.0f;
 			menu->addChild(trackLPFAdjustSlider);
 			
+			StereoWidthLevelSlider *widthSlider = new StereoWidthLevelSlider(&(srcTrack->stereoWidth));
+			widthSlider->box.size.x = 200.0f;
+			menu->addChild(widthSlider);
+			
 			PanCvLevelSlider *panCvSlider = new PanCvLevelSlider(&(srcTrack->panCvLevel));
 			panCvSlider->box.size.x = 200.0f;
 			menu->addChild(panCvSlider);